+++ BackEnd/Areas/Identity/Data/TheMenuBackEndUser.cs
@@ -5,11 +5,15 @@ namespace TheMenu.BackEnd.Areas.Identity.Data;
 // Add profile data for application users by adding properties to the TheMenuBackEndUser class
 public class TheMenuBackEndUser : IdentityUser
 {
-	// TODO Add claims to Identity using IUserClaimsPrincipalFactory<ApplicationUser> https://docs.microsoft.com/en-us/aspnet/core/security/authentication/add-user-data?view=aspnetcore-6.0&tabs=visual-studio#add-claims-to-identity-using-iuserclaimsprincipalfactoryapplicationuser
+    // TODO Add claims to Identity using IUserClaimsPrincipalFactory<ApplicationUser> https://docs.microsoft.com/en-us/aspnet/core/security/authentication/add-user-data?view=aspnetcore-6.0&tabs=visual-studio#add-claims-to-identity-using-iuserclaimsprincipalfactoryapplicationuser
 
     /// <summary>
     ///  TODO Use enum
     /// </summary>
     //public string SocialProvider { get; set; }
+
+    public string? RefreshToken { get; set; }
+
+    public DateTime RefreshTokenExpiryTime { get; set; }  
 }