+++ BackEnd/Migrations/20211003131238_RefreshToken.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using TheMenu.BackEnd.Data;
 
@@ -11,9 +12,10 @@ using TheMenu.BackEnd.Data;
 namespace TheMenu.BackEnd.Migrations
 {
     [DbContext(typeof(TheMenuBackEndContext))]
-    partial class TheMenuBackEndContextModelSnapshot : ModelSnapshot
+    [Migration("20211003131238_RefreshToken")]
+    partial class RefreshToken
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder