+++ BackEnd/Models/AuthResponse.cs
@@ -3,9 +3,15 @@
     public class AuthResponse
     {
         public bool IsAuthSuccessful { get; set; }
+
         public string ErrorMessage { get; set; }
+
         public string Token { get; set; }
-        public bool Is2StepVerificationRequired { get; set; }
+
+        public string RefreshToken { get; set; }
+
+        public DateTime RefreshTokenExpiryTime { get; set; }
+
         public string Provider { get; set; }
     }
 }