+++ BackEnd/Models/TokenApiModel.cs
@@ -1,0 +1,11 @@
+﻿namespace TheMenu.BackEnd.Models
+{
+    public class TokenApiModel
+    {
+        public string AccessToken { get; set; }
+
+        public string RefreshToken { get; set; }
+
+        //TODO Add loginProvider
+    }
+}