+++ FrontEnd/src/app/core/models/auth-response.ts
@@ -2,6 +2,6 @@ export interface AuthResponse {
     isAuthSuccessful: boolean;
     errorMessage: string;
     token: string;
-    is2StepVerificationRequired: boolean;
+    refreshToken: string;
     provider: string;
 }