+++ FrontEnd/src/app/features/landing/landing.module.ts
@@ -3,11 +3,13 @@ import { NgModule } from '@angular/core';
 import { SharedModule } from '../../shared/shared.module';
 import { AboutComponent } from './components/about/about.component';
 import { HomeComponent } from './components/home/home.component';
+import { LoginComponent } from './components/login/login.component';
 
 @NgModule({
     declarations: [
         HomeComponent,
-        AboutComponent
+        AboutComponent,
+        LoginComponent
     ],
     imports: [SharedModule],
     exports: [],