+++ FrontEnd/src/app/features/managers/managers-main/managers-main.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-managers-main',
+  templateUrl: './managers-main.component.html',
+  styleUrls: ['./managers-main.component.scss']
+})
+export class ManagersMainComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}